Added ICHEC import to abc script & added date specification to the top
This should be a more streamlined and appropriate pipeline. I still have to set a working directory for the ICHEC data to be read in, which is not best practice. However, now the date for the ICHEC data is specified at the top, which should prevent over-writing previous iterations. Somewhere I'll need to start keeping a journal of which dates correspond to which set of data when I start wanting to actually use what I'm putting through the pipeline.