Attempt at summary figure of demographics - hard to figure out!
Trying to balance having all of the information presented, and something that makes some kind of mathematical sense. 

For now, different types of experience are stacked with a remaining "other" category that provides an idea of the relative frequency of these experience types, i.e. that there is more experience captured by these questions in the expert group. So, the graph shows each response rate (n=41 for experts & n=123 for non experts) multiplied by 6 (the number of "expertise types" included in the graph.